Rudimentary split of main
